requirements.txt
streamlit
pandas
numpy
plotly-express
scikit-learn
fpdf